pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mesh-bc-tester"
version = "0.1.0"
description = "PySide6 desktop MQTT tester for the Bluetooth Mesh BC light northbound protocol"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.7",
  "paho-mqtt>=2.1",
]

[project.scripts]
mesh-bc-tester = "mesh_bc_tester.main:main"

[tool.setuptools.packages.find]
where = ["src"]
